Shrink CUDA parser generated types and add index_t alias
Switch size_t to int64_t for parser constants (Q, K, table sizes, etc.)
and shrink hash-table span arrays from int64_t to int32_t. Emit a
shared index_t alias (int64_t) for byte-position indices so lexer and
parser agree on the span type. Mirrors the Futhark type changes.

+++ src/Alpacc/Generator/Cuda/Parser.hs
@@ -31,24 +31,24 @@ generateParser parser =
     [i|
 using production_t = #{cudafy production_type};
 using bracket_t = #{cudafy bracket_type};
-const size_t Q = #{cudafy q};
-const size_t K = #{cudafy k};
+const int64_t Q = #{cudafy q};
+const int64_t K = #{cudafy k};
 const terminal_t EMPTY_TERMINAL = #{cudafy $ terminalCast empty_terminal};
 const terminal_t START_TERMINAL = #{cudafy $ terminalCast start_terminal};
 const terminal_t END_TERMINAL = #{cudafy $ terminalCast end_terminal};
-const size_t NUMBER_OF_PRODUCTIONS = #{cudafy number_of_productions};
+const int64_t NUMBER_OF_PRODUCTIONS = #{cudafy number_of_productions};
 __device__ const terminal_t PRODUCTION_TO_TERMINAL[NUMBER_OF_PRODUCTIONS] =
   #{cudafy $ map terminalCast production_to_tertminal};
 __device__ const bool PRODUCTION_TO_TERMINAL_IS_VALID[NUMBER_OF_PRODUCTIONS] =
   #{cudafy production_to_tertminal_is_valid};
-__device__ const size_t PRODUCTION_TO_ARITY[NUMBER_OF_PRODUCTIONS] =
+__device__ const uint8_t PRODUCTION_TO_ARITY[NUMBER_OF_PRODUCTIONS] =
   #{ari};
-const size_t HASH_TABLE_SIZE = #{cudafy $ length $ oaArray oa};
-const size_t MAX_ITERS = #{cudafy $ oaMaxIters oa};
-const size_t PRODUCTIONS_SIZE = #{cudafy productions_size};
-const size_t STACKS_SIZE = #{cudafy stacks_size};
-const size_t MAX_BRACKETS_PER_POSITION = #{cudafy max_brackets_per_pos};
-const size_t MAX_PRODS_PER_POSITION = #{cudafy max_prods_per_pos};
+const int64_t HASH_TABLE_SIZE = #{cudafy $ length $ oaArray oa};
+const int64_t MAX_ITERS = #{cudafy $ oaMaxIters oa};
+const int64_t PRODUCTIONS_SIZE = #{cudafy productions_size};
+const int64_t STACKS_SIZE = #{cudafy stacks_size};
+const int64_t MAX_BRACKETS_PER_POSITION = #{cudafy max_brackets_per_pos};
+const int64_t MAX_PRODS_PER_POSITION = #{cudafy max_prods_per_pos};
 __device__ const bracket_t STACKS[STACKS_SIZE] =
   #{cudafy stacks};
 __device__ const production_t PRODUCTIONS[PRODUCTIONS_SIZE] =
@@ -57,9 +57,9 @@ __device__ const bool HASH_TABLE_IS_VALID[HASH_TABLE_SIZE] =
   #{cudafy hash_table_is_valid};
 __device__ const terminal_t HASH_TABLE_KEYS[HASH_TABLE_SIZE][Q + K] =
   #{cudafy $ map (map terminalCast) hash_table_keys};
-__device__ const int64_t HASH_TABLE_STACKS_SPAN[HASH_TABLE_SIZE][2] =
+__device__ const int32_t HASH_TABLE_STACKS_SPAN[HASH_TABLE_SIZE][2] =
   #{cudafy hash_table_stacks_span};
-__device__ const int64_t HASH_TABLE_PRODUCTIONS_SPAN[HASH_TABLE_SIZE][2] =
+__device__ const int32_t HASH_TABLE_PRODUCTIONS_SPAN[HASH_TABLE_SIZE][2] =
   #{cudafy hash_table_productions_span};
 |]
     <> cudaParser
